Fix Fox drop capture
