+++ arclight-common/src/main/java/io/izzel/arclight/common/mixin/core/world/entity/animal/FoxMixin.java
@@ -1,9 +1,6 @@
 package io.izzel.arclight.common.mixin.core.world.entity.animal;
 
 import io.izzel.arclight.common.bridge.core.entity.passive.FoxEntityBridge;
-import io.izzel.arclight.common.mod.util.ArclightCaptures;
-import io.izzel.arclight.mixin.Decorate;
-import io.izzel.arclight.mixin.DecorationOps;
 import net.minecraft.world.entity.animal.Fox;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.item.ItemStack;
@@ -16,7 +13,6 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import java.util.List;
 import java.util.UUID;
 
 @Mixin(Fox.class)
@@ -34,17 +30,5 @@ public abstract class FoxMixin extends AnimalMixin implements FoxEntityBridge {
     @Inject(method = "pickUpItem", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/item/ItemEntity;discard()V"))
     private void arclight$pickCause(ItemEntity itemEntity, CallbackInfo ci) {
         itemEntity.bridge().bridge$pushEntityRemoveCause(EntityRemoveEvent.Cause.PICKUP);
-    }
-
-    @Decorate(method = "dropAllDeathLoot", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/animal/Fox;spawnAtLocation(Lnet/minecraft/world/item/ItemStack;)Lnet/minecraft/world/entity/item/ItemEntity;"))
-    private ItemEntity arclight$captureFoxDrop(Fox instance, ItemStack itemStack) throws Throwable {
-        try {
-            arclight$spawnNoAdd = true;
-            final var result = (ItemEntity) DecorationOps.callsite().invoke(instance, itemStack);
-            ArclightCaptures.captureExtraDrops(List.of(result));
-            return result;
-        } finally {
-            arclight$spawnNoAdd = false;
-        }
     }
 }
