+++ arclight-fabric/src/main/resources/mixins.arclight.fabric.json
@@ -23,6 +23,7 @@
     "core.network.ServerLoginPacketListenerImplMixin_Fabric",
     "core.world.entity.EntityMixin_Fabric",
     "core.world.entity.LivingEntityMixin_Fabric",
+    "core.world.entity.animal.FoxMixin_Fabric",
     "core.world.entity.animal.SheepMixin_Fabric",
     "core.world.entity.animal.WolfMixin_Fabric",
     "core.world.entity.monster.ZombieMixin_Fabric",
