+++ arclight-forge/src/main/resources/mixins.arclight.forge.json
@@ -23,9 +23,9 @@
     "core.command.CommandsMixin_Forge",
     "core.fluid.LavaFluidMixin_Forge",
     "core.network.ServerCommonPacketListenerImplMixin_Forge",
+    "core.network.ServerGamePacketListenerImplMixin_Forge",
     "core.network.ServerHandshakePacketListenerImplMixin_Forge",
     "core.network.ServerLoginPacketListenerImplMixin_Forge",
-    "core.network.ServerGamePacketListenerImplMixin_Forge",
     "core.network.ServerStatusNetHandlerMixin_Forge",
     "core.network.protocol.login.ServerboundCustomQueryAnswerPacketMixin_Forge",
     "core.network.protocol.status.ServerStatusMixin",
@@ -38,6 +38,7 @@
     "core.world.entity.LivingEntityMixin_Forge",
     "core.world.entity.MobMixin_Forge",
     "core.world.entity.ai.behavior.HarvestFarmlandMixin_Forge",
+    "core.world.entity.animal.FoxMixin_Forge",
     "core.world.entity.animal.MushroomCowMixin_Forge",
     "core.world.entity.animal.SheepMixin_Forge",
     "core.world.entity.animal.WolfMixin_Forge",
